requirements.txt
tensorflow-gpu
pandas
scikit-learn